+++ instrumentation/finagle-http-23.11/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/v23_11/FinagleCoreInstrumentationModule.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class FinagleCoreInstrumentationModule extends InstrumentationModule {
 
   public FinagleCoreInstrumentationModule() {
-    super("finagle-http");
+    super("finagle", "finagle-23.11", "finagle-http", "finagle-http-23.11");
   }
 
   @Override