+++ instrumentation/spring/spring-boot-autoconfigure/src/test/java/io/opentelemetry/instrumentation/spring/autoconfigure/resources/SpringConfigPropertiesTest.java
@@ -33,6 +33,7 @@ class SpringConfigPropertiesTest {
                       env,
                       new SpelExpressionParser(),
                       new OtlpExporterProperties(),
+                      new OtelResourceProperties(),
                       new PropagationProperties());
 
               assertThat(config.getMap("otel.springboot.test.map"))