+++ smoke-tests-otel-starter/src/test/java/io/opentelemetry/smoketest/OtelSpringStarterDisabledSmokeTest.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.smoketest;
+
+import io.opentelemetry.spring.smoketest.OtelSpringStarterSmokeTestApplication;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.DisabledInNativeImage;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(
+    classes = {
+      OtelSpringStarterSmokeTestApplication.class,
+      OtelSpringStarterSmokeTest.TestConfiguration.class
+    },
+    properties = {"otel.sdk.disabled=true"})
+@DisabledInNativeImage // Without this the native tests in the OtelSpringStarterSmokeTest class will
+// fail with org.h2.jdbc.JdbcSQLSyntaxErrorException: Table "TEST_TABLE"
+// already exists
+class OtelSpringStarterDisabledSmokeTest {
+
+  @Test
+  void shouldStartApplication() {
+    // make sure we can still start the application with the disabled property
+  }
+}